requirements_dev.txt
alabaster==0.7.10
appdirs==1.4.3
arrow==0.10.0
Babel==2.4.0
binaryornot==0.4.3
certifi==2017.4.17
chardet==3.0.3
click==6.7
cookiecutter==1.5.1
docutils==0.13.1
funcsigs==1.0.2
future==0.16.0
idna==2.5
imagesize==0.7.1
Jinja2==2.9.6
jinja2-time==0.2.0
MarkupSafe==1.0
mock==2.0.0
packaging==16.8
pbr==3.0.1
poyo==0.4.1
psycopg2==2.7.1
py==1.4.33
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.1.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.17.3
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.2
sphinxcontrib-websupport==1.0.1
typing==3.6.1
urllib3==1.21.1
whichcraft==0.4.1
